chore: Group guards and fix file paths
