+++ frontend/src/hooks/guards/useAuthenticated.ts
